requirements.txt
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.10.4
autopep8==2.3.2
black==26.3.1
certifi==2026.2.25
click==8.3.3
deprecation==2.1.0
et_xmlfile==2.0.0
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.0
hyperframe==6.1.0
idna==3.12
mypy_extensions==1.1.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.1
pandas==3.0.2
pathspec==1.1.0
platformdirs==4.9.6
postgrest==0.18.0
pycodestyle==2.14.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytokens==0.4.1
pytz==2026.1.post1
realtime==2.28.3
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
strip-hints==0.1.13
supabase==2.10.0
supafunc==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
websockets==15.0.1
wheel==0.47.0
